A beginning is a very delicate time.
Know then that it is the year 10,191.

The Known Universe is ruled by the Padishah Emperor Shaddam IV, my father.
In this time, the most precious substance in the universe is the spice Melange.
The spice extends life. The spice expands consciousness.
The spice is vital to space travel. The Spacing Guild and its navigators,
who the spice has mutated over 4,000 years, use the orange spice gas,
which gives them the ability to fold space. That is, travel to any part of the
universe without moving. Oh, yes. I forgot to tell you - the spice exists on
only one planet in the entire universe. A desolate, dry planet with vast
deserts. Hidden away within the rocks of these deserts are a people known as
the Fremen, who have long held a prophecy that a man would come,
a messiah who would lead them to true freedom. The planet is Arrakis,
also known as Dune.