nixpkcs.keypairs.*.id: support huge key IDs
Some tokens (the nitrokey) actually use an ASCII string for the key ID.
This isn't very common (both the yubikey and TPM2 use positive integers)
but we should be able to support it by passing a very large number to
the `id` value and converting it to URL notation as a big-endian
integer.

Update tests for this case.